+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ================
 * [#1387](https://github.com/oshi/oshi/pull/1387): Switch tests to JUnit5 and Hamcrest matchers - [@dbwiddis](https://github.com/dbwiddis).
 * [#1388](https://github.com/oshi/oshi/pull/1388): Support systemctl list-unit-files return 3 and more columns - [@Szwendacz99](https://github.com/Szwendacz99).
+* [#1389](https://github.com/oshi/oshi/pull/1389): Vendor frequency from non-Intel chips - [@dbwiddis](https://github.com/dbwiddis).
 * Your contribution here
 
 5.3.0 (2020-10-11), 5.3.1 (2020-10-18), 5.3.2 (2020-10-25), 5.3.3 (2020-10-28)